+++ gantry/commands/__init__.py
@@ -1,8 +1,9 @@
 from .cluster import cluster
 from .config import config
 from .follow import follow
+from .list import list_cmd
 from .main import main
 from .run import run
 from .stop import stop
 
-__all__ = ["main", "follow", "run", "config", "cluster", "stop"]
+__all__ = ["main", "follow", "run", "config", "cluster", "stop", "list_cmd"]